Load IPython configuration into ipdb
Now load a full IPythonApp in the background in order to retrieve the
configuration and inject it into ipdb.

If an instance of IPython is already running its configuration will be loaded.
If it is an embed shell and no IPythonApp is running, the logic is usually in a
python script and you will have to change the code in order to make it works.