Add minimal PEP517 setup
This commit avoids use of the deprecated `setup_requires` to replace the
currently-incorrect use of `install_requires` by setting up a minimal
PEP517 compatible `pyproject.toml`. `install_requires` is for runtime
dependencies while the PEP517 `build-system` stuff is where you specify
what is needed to build or install the project, not what you need to
import the module.